requirements.txt
redis
collections
mysql-connector-python
flask